+++ client/client.go
@@ -97,7 +97,6 @@ func (lr *loggingRT) RoundTrip(req *http.Request) (res *http.Response, err error
 	defer func() {
 		l := lr.l.WithContext(ctx)
 		msg := "http_client"
-		// TODO: (komuw), check if using []any{} is okay.
 		flds := []any{
 			"method", req.Method,
 			"url", req.URL.Redacted(),
@@ -116,7 +115,6 @@ func (lr *loggingRT) RoundTrip(req *http.Request) (res *http.Response, err error
 		}
 	}()
 
-	// TODO: (komuw) should we use l.logID if id from `log.GetId` did not come from ctx.
 	id, _ := log.GetId(ctx)
 	req.Header.Set(logIDHeader, id)